MooringPoint

The MooringPoint module manages individual points in the mooring system, such as anchors, fairleads, and connection points.

Types

Mooring.MooringPoints.MooringPoint — Type

MooringPoint Struct This struct is used to define a point in the mooring system. It includes the following fields:

tag::String: Identifier for the point

btrian::Triangulation: BoundaryTriangulation or InterfaceTriangulation with the position of the point in space

motion::MotionType: Type of motion of the point. If no motion is specified, the point is free to move in space.

Functions

Mooring.MooringPoints.get_triangulation — Function

get_triangulation(p::MooringPoint) Get the boundary/interface triangulation of a point

Additional functions:

get_motion: Returns the motion type of the point

get_boundary_triangulation: Returns the boundary triangulation

get_measures: Returns measure for integration