Drag force

Mooring.Drag.ChainStudless — Type

ChainStudless Struct

This type is used to define default drag properties for a studless chain.

Mooring.Drag.ChainStudlink — Type

ChainStudlink Struct

This type is used to define default drag properties for a studlink chain.

Mooring.Drag.Custom — Type

Custom Struct

This type is used to define custom drag properties.

Mooring.Drag.DragProperties — Type

DragProperties Struct

This struct contains the properties of the drag. The following parameters are included, no default values are assumed.

dragType::DragType: Type of drag

ρw::Real: Water density [Kg/m3]

nd::Real: Nominal diameter [m]

od::Real: Outer diameter [m]

id::Real: Inner diameter [m]

AStr::Real: Area (structural) [m2]

Cd_n::Real: Drag coefficient normal

Cd_t::Real: Drag coefficient tangent

dd_n::Real: Drag diameter normal

dd_t::Real: Drag diameter tangent

Ca_n::Real: Added mass coefficient normal

Ca_t::Real: Added mass coefficient tangent

Cfd_n::Real: Friction coefficient normal

Cfd_t::Real: Friction coefficient tangent

Mooring.Drag.DragProperties — Method

DragProperties Constructor

This function constructs the drag properties for the ChainStudless type.

Input:

dragType::ChainStudless: Type of drag

nd::Real: Nominal diameter [m]

AStr::Real: Area (structural) [m2]

Optional parameters (with default values):

Cd_n::Real = 2.4: Drag coefficient normal

Cd_t::Real = 1.15: Drag coefficient tangent

Ca_n::Real = 1.0: Added mass coefficient

Ca_t::Real = 0.5: Added mass coefficient

ρw::Real = 1025: Water density [Kg/m3]

od::Real = 1.80 * nd: Outer diameter [m]

id::Real = 0.0: Inner diameter [m]

dd_n::Real = nd: Drag diameter normal

dd_t::Real = nd / π: Drag diameter tangent

Output:

DragProperties: Drag properties

Mooring.Drag.DragProperties — Method

DragProperties Constructor

This function constructs the drag properties for the ChainStudlink type.

Input:

dragType::ChainStudlink: Type of drag

nd::Real: Nominal diameter [m]

AStr::Real: Area (structural) [m2]

Optional parameters (with default values):

Cd_n::Real = 2.6: Drag coefficient normal

Cd_t::Real = 1.4: Drag coefficient tangent

Ca_n::Real = 1.0: Added mass coefficient

Ca_t::Real = 0.5: Added mass coefficient

ρw::Real = 1025: Water density [Kg/m3]

od::Real = 1.89 * nd: Outer diameter [m]

id::Real = 0.0: Inner diameter [m]

dd_n::Real = nd: Drag diameter normal

dd_t::Real = nd / π: Drag diameter tangent

Mooring.Drag.DragProperties — Method

DragProperties Constructor

This function constructs the drag properties for the Custom type.

Input:

dragType::Custom: Type of drag

nd::Real: Nominal diameter [m]

AStr::Real: Area (structural) [m2]

Optional parameters (with default values):

Cd_n::Real: Drag coefficient normal

Cd_t::Real: Drag coefficient tangent

Ca_n::Real: Added mass coefficient

Ca_t::Real: Added mass coefficient

ρw::Real = 1025: Water density [Kg/m3]

od::Real = nd: Outer diameter [m]

id::Real = 0.0: Inner diameter [m]

dd_n::Real = nd: Drag diameter normal

dd_t::Real = nd / π: Drag diameter tangent

Mooring.Drag.DragProperties — Method

DragProperties Constructor

This function constructs the drag properties for the NoDrag type.

Input:

dragType::NoDrag: Type of drag

Output:

DragProperties: Drag properties

Mooring.Drag.DragType — Type

DragType Struct

This abstract type is used to define the type of drag. Possible implemented options are:

NoDrag: No drag

Custom: Custom drag

ChainStudless: Studless chain drag

ChainStudlink: Studlink chain drag

Mooring.Drag.NoDrag — Type

NoDrag Struct

Using this type all drag properties are set to zero.

Mooring.Drag.drag_ΓX — Method

drag_ΓX

This function calculates the drag force on the line. The drag force is given by:

\[F_{\text{drag}} = (C_{\text{fd}_n} \cdot v_n \cdot \|v_n\| + C_{\text{fd}_t} \cdot v_t \cdot \|v_t\|) \cdot s_\Lambda\]

Input:

dragProp::DragProperties: Drag properties

sΛ::Real: line stretch

t::VectorValue: Line tangent

vr::VectorValue: Relative velocity

Output:

VectorValue: Drag force

Mooring.Drag.relative_velocity — Method

relative_velocity

This function calculates the relative velocity between the line and the fluid. This implementation computes the difference between the line ($ v $) and the fluid ($ u $) velocities, $ v_{\text{rel}} = u - v $.

Input:

u::VectorValue: Fluid velocity

v::VectorValue: Line velocity

Output:

VectorValue: Relative velocity

Mooring.Drag.relative_velocity — Method

relative_velocity

This function calculates the relative velocity between the line and the fluid. The default implementation is the negative of the line velocity, $ v_{\text{rel}} = -v $.

Input:

v::VectorValue: Line velocity

Output:

VectorValue: Relative velocity

Mooring.Drag.relative_velocity — Method

relative_velocity

This function calculates the relative velocity between the line and the fluid. This implementation computes the difference between the line ($ v $) and the fluid ($ u $) velocities, assuming a time ramp function for the fluid velocity ($ \alpha(t) $). The relative velocity is given by $ v_{\text{rel}} = \alpha(t)*u - v $.

Input:

t::Real: Time

input_ramp::WTS.TimeRampTypeAll: Time ramp function for the fluid velocity

u::VectorValue: Fluid velocity

v::VectorValue: Line velocity

Output:

VectorValue: Relative velocity