Environmental Conditions

Mooring.EnvironmentalConditions.get_current_field — Method

getcurrentfield

This function returns the current field at a given point of the undeformed configuration $ r \rightarrow X_h(r) $.

Input:

r::Real: Point in the undeformed configuration

Xh::Function: Function that maps the undeformed configuration to the deformed configuration (VectorValue)

curObj::CurrentStat: Current object

Output:

VectorValue: Current field at the point r

Mooring.EnvironmentalConditions.get_input_spectrum — Method

getinputspectrum

This function returns the wave spectrum parameters for a given set of wave parameters. The wave spectrum is calculated using the package WaveSpec.

Mooring.EnvironmentalConditions.get_wave_velocity — Method

getwavevelocity

This function returns the wave velocity at a given point in the space x. It uses the Airy wave theory, defined in the WaveSpec package.

Input:

t::Real: Time

sp::SpecStruct: Wave spectrum parameters

x::VectorValue: Point in the space

Output:

VectorValue: Wave velocity at the point x