Point Motion

Mooring.PointMotion.CustomMotionType — Type

CustomMotion Struct

This struct is used to define the properties of the custom motion. This requires a function of space and time. - f::Function: custom functions

Mooring.PointMotion.MotionType — Type

MotionType Struct

This abstract type is used to define the type of motion. Possible implemented options are:

WaveMotion: Motion given by a waves (regular or irregular)

CustomMotion: Custom function

Mooring.PointMotion.MotionType — Method

MotionType constructor

This function returns the motion of a point given its parameters. Input:

p_params::PointParameters: Parameters of the point

Output:

motion::MotionType: Motion of the point

Mooring.PointMotion.WaveMotionType — Type

WaveMotion Struct

This struct is used to define the properties of the wave motion. This includes: - waveparams::WaveParameters: wave Parameters - pointmotion_function::Function: function that returns the position of a point in time

This struct precomputes the motion of a point given the motion type, input ramp, time and position. For a WaveMotionType the motion is calculated using the wave spectrum and the Airy wave theory. It calculates the position of a point in time at certain given time instances (t_range) and constructs an interpolable function that returns the interpolated value at any given time.

This pre-computation is done to avoid evaluating the wave spectrum at each time step.

Mooring.PointMotion.get_point_motion_function — Method

getpointmotion This function returns the motion function of a point given its motion type. Input:

motion_type::MotionType: Type of motion (WaveMotionType or CustomMotionType)

Output:

point_motion_function::Function: Interpolated position vector. It can be evaluated at any time instance