Sea bed external loading

Mooring.SeaBed.ramp_linear — Method

ramp_linear

This function returns the linear ramp function for a given excursion into the sea bed. It assumes the excursion is positive when there is penetration into the soil and computes the value based on the penetration_depth_ramp parameter:

\[x_{\text{new}} = \frac{x}{\text{penetration_depth_ramp}}.\]

Input:

params::SeaBedParameters: Sea bed parameters

exc::Real: excursion value at a given time

Output:

Real: Linear ramp function value

Mooring.SeaBed.ramp_tanh — Method

ramp_tanh

This function returns the tanh ramp function for a given excursion into the sea bed. It assumes the excursion is positive when there is penetration to the sea bed and negative when the line is lifted from the soil. The function is defined as:

\[x_{\text{new}} = \max(0, 2 \tanh( \text{tanh_ramp} \cdot x ) ).\]

Input:

params::SeaBedParameters: Sea bed parameters

exc::Real: excursion value at a given time

Output:

Real: Tanh ramp function value

Mooring.SeaBed.sea_bed_force — Method

seabedforce

This function computes the force exerted by the sea bed on the mooring line. It considers the excursion of the line into the sea bed, the velocity of the line, and the properties of the sea bed. The function is defined as:

\[\alpha_1 * w + \alpha_2 * k * \Lambda * ( x_z - \beta_1 * v_z - \beta_2 * v_z * \abs(v_z) )\]

with $ \Lambda $ the stretch along the line, $ xz $ the vertical excursion, and $ vz $ the vertical velocity. Computed as:

\[x_z = \mathbf{e}_z \cdot (\mathbf{X} + \mathbf{u}) V_z = \mathbf{e}_z \cdot \mathbf{v}\]

Input:

params::SeaBedParameters: Sea bed parameters

X::VectorValue: Position of the line

u::VectorValue: Displacement of the line

v::VectorValue: Velocity of the line

Λ::Real: Stretch along the line

e_z::VectorValue: Unit vector in the vertical direction

Output:

VectorValue: Force exerted by the sea bed at a given point in the line

Mooring.SeaBed.set_still_weight — Method

setstillweight

This function sets the still weight of the sea bed. It assumes the sea bed object is already created and modifies the still weight parameter.

Input:

params::SeaBedParameters: Sea bed parameters

still_weight::Real: Still weight [N]

Output:

SeaBedParameters: Sea bed parameters with the still weight modified