Tangential Differential Calculus

Mooring.TangentialDiffCalculus.Edir — Method

Edir(FΓ::TensorValue)

Directional Green-Lagrange strain

This function returns the directional Green-Lagrange strain for a given line deformation gradient FΓ. The directional Green-Lagrange strain is defined as:

\[\mathbf{E}_{\text{dir}} = 0.5(\mathbf{F}_Γ^T⋅\mathbf{F}_Γ - \mathbf{I})\]

where I is the identity tensor. The dimensions of the directional Green-Lagrange strain are n×n, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.Etang — Method

Etang(P::TensorValue,Edir::TensorValue)

Tangential Green-Lagrange strain

This function returns the tangential Green-Lagrange strain for a given projection operator P and directional Green-Lagrange strain Edir. The tangential Green-Lagrange strain is defined as:

\[\mathbf{E}_{\text{tang}} = \mathbf{P}⋅\mathbf{E}_{\text{dir}}⋅\mathbf{P}\]

The dimensions of the tangential Green-Lagrange strain are n×n, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.FΓ — Method

FΓ(u::CellField,X::CellField)

Line deformation gradient

This function returns the line deformation gradient for a given displacement field u, and a map field X. The line deformation gradient is defined as:

\[F_Γ = ∇_{x}^{Γ,dir}(u) + I\]

where I is the identity tensor. The dimensions of the line deformation gradient are n×n, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.FΓ — Method

FΓ(∇xΓdir_u::TensorValue)

∇xΓdir (Line deformation gradient)

Mooring.TangentialDiffCalculus.G — Method

G(J)

Metric tensor

This function returns the metric tensor of a given coordinate map field X. The metric tensor is defined as: $ \mathbf{G} = \mathbf{J}^T \cdot \mathbf{J} $, with $ \mathbf{J} $ the J operator of the map field X.

Note that the dimensions of the metric tensor are 1×1.

Mooring.TangentialDiffCalculus.J — Method

J(X::CellField)

Jacobian operator

This function returns the Jacobian operator of a given coordinate map field X. We use the transpose as the gradient operator is defined as ∇(X)=∇×X in Gridap. The Jacobian operator is defined as:

\[mathbf{J}(mathbf{X}) = ∇_r(mathbf{X}) = rac{∂X_i}{∂r_j}\]

Note that the dimensions of the Jacobian operator are n×1, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.P — Method

P(J)

Projection operator

This function returns the projection operator of a given vector field u onto the tangent space of the map field X. The projection operator is defined as:

\[\mathbf{P} = \frac{ \mathbf{J}⋅\mathbf{J} }{ \|\mathbf{J}\|}^2}}\]

Note that the dimensions of the projection operator are n×n, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.Q — Method

Q(J)

Transformation matrix

This function returns the transformation matrix $ \mathbf{Q} $ of a given coordinate map field X. The transformation matrix is defined as:

\[\mathbf{Q} = \mathbf{J}\cdot\mathbf{G}^{-1}\]

Note that the dimensions of the transformation matrix are n×1, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.g — Method

g(j)

Metric tensor in the physical space

This function returns the metric tensor in the physical space as:

\[g = j^T⋅j\]

Note that the dimensions of the metric tensor in the physical space are 1×1.

Mooring.TangentialDiffCalculus.j — Method

j(FΓ,J)

Jacobian operator in the physical space

This function returns the Jacobian determinant in the physical space for a given line deformation gradient FΓ and Jacobian operator J. The Jacobian determinant in the physical space is defined as:

\[j = FΓ^T⋅J\]

Note that the dimensions of the Jacobian determinant in the physical space are n×1.

Mooring.TangentialDiffCalculus.q — Method

q(j)

Transformation matrix in the physical space

This function returns the transformation matrix in the physical space as:

\[q = j⋅g^{-1}\]

Note that the dimensions of the transformation matrix in the physical space are n×1, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.Λ — Method

Λ(j,J)

Line stretch

This function returns the stretch of a given line. The stretch is defined as:

\[Λ = \frac{\det(g)}{\det(G)}\]

where g is the metric tensor in the physical space, and G is the metric tensor in the reference space. The result is a scalar.

Mooring.TangentialDiffCalculus.∇ₓΓdir — Method

∇ₓΓdir(u::CellField,X::CellField)

Tangential gradient

This function returns the tangential gradient of a given vector field u. The tangential gradient is defined as:

\[∇_{x}^{Γ,dir}(u) = ∇_r(u)⋅Q\]

Note that the dimensions of the tangential gradient are n×n, where n is the dimension of the physical space.

Mooring.TangentialDiffCalculus.∇ₓΓdir — Method

∇ₓΓdir(∇u::TensorValue,J::TensorValue)

∇ₓΓdir (Tangential gradient)